+++ Sources/VCore/Extensions/Foundation/Array.FirstIndexAndElement.swift
@@ -13,14 +13,14 @@ extension Array {
     ///
     /// Used for avoiding accessing both index and element with built-in methods.
     ///
-    /// Usage Example
+    /// Usage Example:
     ///
     ///     let nums: [Int] = [1, 3, 5]
     ///     guard let (index, num): (Int, Int) = nums.firstIndexAndElement(where: { $0 * $0 >= 10 }) else { return }
     ///
     public func firstIndexAndElement(
         where predicate: (Element) throws -> Bool
-    ) rethrows -> (Int, Element)? {
+    ) rethrows -> (index: Int, element: Element)? {
         guard let index: Int = try firstIndex(where: predicate) else { return nil }
         let element: Element = self[index]