+++ Sources/VCore/Extensions/Foundation/FloatingPoint.ClampedInRange.swift
@@ -13,25 +13,6 @@ extension FloatingPoint {
     ///
     /// Usage Example:
     ///
-    ///     var value1: Double = 0.0; value1.clamp(in: 1...10) // 1.0
-    ///     var value2: Double = 5.0; value2.clamp(in: 1...10) // 5.0
-    ///     var value3: Double = 11.0; value3.clamp(in: 1...10) // 10.0
-    ///
-    ///     var value4: Double = 0.0; value4.clamp(in: 1...10, step: 3) // 1.0
-    ///     var value5: Double = 5.0; value5.clamp(in: 1...10, step: 3) // 4.0
-    ///     var value6: Double = 11.0; value6.clamp(in: 1...10, step: 3) // 10.0
-    ///
-    public mutating func clamp(
-        in range: ClosedRange<Self>,
-        step: Self? = nil
-    ) {
-        self = self.clamped(in: range, step: step)
-    }
-    
-    /// Clamps value in range.
-    ///
-    /// Usage Example:
-    ///
     ///     let value1: Double = 0.0.clamped(in: 1...10) // 1.0
     ///     let value2: Double = 5.0.clamped(in: 1...10) // 5.0
     ///     let value3: Double = 11.0.clamped(in: 1...10) // 10.0
@@ -52,7 +33,7 @@ extension FloatingPoint {
         }
     }
     
-    /// Clamps value between `min` and `max` values.
+    /// Clamps value in range.
     ///
     /// Usage Example:
     ///
@@ -64,12 +45,11 @@ extension FloatingPoint {
     ///     var value5: Double = 5.0; value5.clamp(in: 1...10, step: 3) // 4.0
     ///     var value6: Double = 11.0; value6.clamp(in: 1...10, step: 3) // 10.0
     ///
-    public mutating func clamp(
-        min: Self,
-        max: Self,
+    mutating public func clamp(
+        in range: ClosedRange<Self>,
         step: Self? = nil
     ) {
-        self = self.clamped(min: min, max: max, step: step)
+        self = self.clamped(in: range, step: step)
     }
     
     /// Returns value clamped between `min` and `max` values.
@@ -90,5 +70,25 @@ extension FloatingPoint {
         step: Self? = nil
     ) -> Self {
         clamped(in: min...max, step: step)
+    }
+    
+    /// Clamps value between `min` and `max` values.
+    ///
+    /// Usage Example:
+    ///
+    ///     var value1: Double = 0.0; value1.clamp(min: 1, max: 10) // 1.0
+    ///     var value2: Double = 5.0; value2.clamp(min: 1, max: 10) // 5.0
+    ///     var value3: Double = 11.0; value3.clamp(min: 1, max: 10) // 10.0
+    ///
+    ///     var value4: Double = 0.0; value4.clamp(min: 1, max: 10, step: 3) // 1.0
+    ///     var value5: Double = 5.0; value5.clamp(min: 1, max: 10, step: 3) // 4.0
+    ///     var value6: Double = 11.0; value6.clamp(min: 1, max: 10, step: 3) // 10.0
+    ///
+    mutating public func clamp(
+        min: Self,
+        max: Self,
+        step: Self? = nil
+    ) {
+        self = self.clamped(min: min, max: max, step: step)
     }
 }