+++ Sources/VCore/Extensions/Foundation/String.KeepingOnlyChars.swift
@@ -9,28 +9,28 @@ import Foundation
 
 // MARK: - String Keeping Only Chars
 extension String {
-    /// Filters out characters that don't match the criteria.
-    ///
-    /// Usage Example:
-    ///
-    ///     var phoneNumber = "+0123456789"
-    ///     phoneNumber.keep(only: .decimalDigits)
-    ///
-    mutating public func keep(only characterSet: CharacterSet) {
-        self = self.keeping(only: characterSet)
-    }
-    
     /// Returns string with chracters, that don't match the criteria, filtered out.
     ///
     /// Usage Example:
     ///
-    ///     let phoneNumber = "+0123456789"
-    ///     let phoneNumberStr: String = phoneNumber.keeping(only: .decimalDigits)
+    ///     let phoneNumber: String = "+0123456789"
+    ///     let phoneNumber2: String = phoneNumber.keeping(only: .decimalDigits) // "0123456789"
     ///
     public func keeping(only characterSet: CharacterSet) -> String {
         filter { char in
             guard let unicodeScalar = char.unicodeScalars.first else { return false }
             return characterSet.contains(unicodeScalar)
         }
+    }
+    
+    /// Filters out characters that don't match the criteria.
+    ///
+    /// Usage Example:
+    ///
+    ///     var phoneNumber: String = "+0123456789"
+    ///     phoneNumber.keep(only: .decimalDigits) // "0123456789"
+    ///
+    mutating public func keep(only characterSet: CharacterSet) {
+        self = self.keeping(only: characterSet)
     }
 }