+++ Sources/VCore/Services and Managers/Network Client/Network Request Factory/NetworkRequestFactory.QueryParameters.swift
@@ -22,7 +22,7 @@ extension NetworkRequestFactory {
             for (key, value) in json {
                 guard let value = value else { continue }
                 
-                guard let description: String = .init(safelyDescribing: value) else { throw NetworkError.invalidQueryparameters }
+                guard let description: String = .init(unwrappedDescribing: value) else { throw NetworkError.invalidQueryparameters }
                 result.updateValue(description, forKey: key)
             }
             
@@ -39,7 +39,7 @@ extension NetworkRequestFactory {
             for (key, value) in json {
                 guard let value = value else { continue }
                 
-                guard let description: String = .init(safelyDescribing: value) else { throw NetworkError.invalidQueryparameters }
+                guard let description: String = .init(unwrappedDescribing: value) else { throw NetworkError.invalidQueryparameters }
                 result.updateValue(description, forKey: key)
             }